+++ CHANGELOG.md
@@ -15,6 +15,12 @@ All notable changes to this project will be documented here.
 
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Bug fixes
+
+- Cancel pending callback after unmounting a component in `useAsyncCallback`.
+
 ## 1.0.0-rc.3 (August 20, 2019)
 
 ### Bug fixes